requirements.txt
aiohttp==3.7.4.post0
py-cord==1.7.3
discord-py-slash-command==3.0.1
python-decouple==3.4